Extract block domain to folder
Keep all of the block-related code in one place.
