+++ client/actions/index.ts
@@ -1,7 +1,6 @@
 export * from './accounts';
 export * from './ajax';
 export * from './author';
-export * from './block';
 export * from './commits';
 export * from './editor';
 export * from './form';
