+++ client/block/index.ts
@@ -1,0 +1,4 @@
+import * as actions from './actions';
+import * as config from './config';
+
+export { actions, config };
