+++ client/components/index.ts
@@ -1,8 +1,8 @@
 export { default as Blob } from './Blob';
-export * from './Block';
 export { default as Commits } from './Commits';
 export { default as EditPage } from './EditPage';
 export { default as SearchPopup } from './SearchPopup';
 export { default as SettingsPage } from './SettingsPage';
 export * from './Shortcode';
+export * from './StatusMapper';
 export { default as TextControl } from './TextControl';
