+++ client/util.ts
@@ -1,7 +1,5 @@
 import { ActionType } from 'typesafe-actions';
 import * as t from 'io-ts';
-import * as actions from './actions';
-import { actions as searchActions } from './search';
 
 // @todo fix this type
 export type Loopable<I extends string, E> = {
@@ -17,4 +15,8 @@ export type Toggle = t.TypeOf<typeof toggle>;
 
 export type Cursor = false | [number, number];
 
-export type RootAction = ActionType<typeof actions & typeof searchActions>;
+export type RootAction = ActionType<
+  typeof import('./actions') &
+    typeof import('./search').actions &
+    typeof import('./block').actions
+>;
